Add exceptions that don't print a stack
